+++ projects/np-ui-lib/src/lib/np-sidepanel/np-sidepanel.module.ts
@@ -1,14 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpSidepanelComponent } from './np-sidepanel.component';
-import { NpSidepanelHeaderComponent } from './np-sidepanel-header.component';
-import { NpSidepanelFooterComponent } from './np-sidepanel-footer.component';
 
 @NgModule({
-  declarations: [NpSidepanelComponent, NpSidepanelHeaderComponent, NpSidepanelFooterComponent],
+  declarations: [NpSidepanelComponent],
   imports: [
     CommonModule
   ],
-  exports: [NpSidepanelComponent, NpSidepanelHeaderComponent, NpSidepanelFooterComponent]
+  exports: [NpSidepanelComponent]
 })
 export class NpSidepanelModule { }