+++ projects/np-ui-lib/src/public-api.ts
@@ -118,8 +118,6 @@ export * from './lib/np-blockui/np-blockui.directive';
 /*Sidepanel */
 export * from './lib/np-sidepanel/np-sidepanel.module';
 export * from './lib/np-sidepanel/np-sidepanel.component';
-export * from './lib/np-sidepanel/np-sidepanel-header.component';
-export * from './lib/np-sidepanel/np-sidepanel-footer.component';
 
 /*Breadcrumb */
 export * from './lib/np-breadcrumb/np-breadcrumb.module';