+++ src/app/np-popover-demo/np-popover-demo.component.ts
@@ -9,7 +9,9 @@ export class NpPopoverDemoComponent implements OnInit {
 
   importText = 'import { NpPopoverModule } from \'np-ui-lib\';';
   htmlText = `<a np-popover [header]="'Link header'" [body]="'Link Description.'">Link</a>`;
-  contextText = `<a np-popover [header]="headerTemp" [body]="bodyTemp" [context]="{count: 5}">Context Sample</a>
+  contextText = `<a np-popover [header]="headerTemp" [body]="bodyTemp" [context]="{count: 5}">
+  Context Sample
+</a>
 <ng-template #headerTemp let-count="count">
   Count pass in context is {{count}}.
 </ng-template>`;