+++ src/app/np-utility-demo/np-utility-demo.component.ts
@@ -12,7 +12,15 @@ export class NpUtilityDemoComponent implements OnInit {
   text = 'A paragraph is a series of related sentences developing a central idea, called the topic.Try to think about paragraphs in terms of thematic unity: a paragraph is a sentence or a group of sentences that supports one central, unified idea.Paragraphs add one idea at a time to your broader argument.';
   highlightTextCode = `<p [innerHTML]="text | npHighLight: 'paragraph'"></p>`;
   autofocusTextCode = `<input type="text" [npAutoFocus]="true">`;
-
+  toolbarCode = `<div class="np-toolbar">
+  <div class="np-toolbar-item">...</div>
+  <div class="np-toolbar-item">...</div>
+  <div class="np-toolbar-separator"></div>
+  ...
+  <div class="np-toolbar-spacer"></div>
+  <div class="np-toolbar-item">...</div>
+  <div class="np-toolbar-item">...</div>
+</div>`
   firstName: string;
   lastName: string;
   birthDate: Date;