+++ projects/np-ui-lib/CHANGELOG.md
@@ -6,6 +6,7 @@
 ## 10.8.4
 - Minor improvements
 - Add font and size in rich text editor
+- Recreate sidepanel component
 
 ## 10.8.3
 - Add theme and scss files which were missing in 10.8.2